+++ config/eslintrc/src/rules/syntax/base.ts
@@ -25,7 +25,11 @@ import { defineFlatConfig } from "eslint-flat-config-utils";
 
 const files = (() => filesTS(["**/*.js", "**/*.jsx", "**/*.mjs", "**/*.ejs"]))();
 const languageOptions = (() =>
-	languageOptionsTS({ ecmaVersion: "latest", sourceType: "module", globals: { ...globals.node } }))();
+	languageOptionsTS({
+		ecmaVersion: "latest",
+		sourceType: "module",
+		globals: { ...globals.node }
+	}))();
 
 export default defineFlatConfig({
 	files: files,
@@ -114,7 +118,7 @@ export default defineFlatConfig({
 		"no-unused-vars": "warn",
 		"class-methods-use-this": "error",
 		"no-use-before-define": "off",
-		complexity: ["error", 10],
+		complexity: ["error", 16],
 		"no-useless-backreference": "error",
 		// "consistent-return": "error",
 		"require-atomic-updates": "warn",
@@ -265,7 +269,7 @@ export default defineFlatConfig({
 		"@EslintOptRegConf/optimize-regex": "error",
 		"@EslintSecurity/detect-unsafe-regex": "error",
 
-		"@EslintSecurity/detect-object-injection": "error",
+		"@EslintSecurity/detect-object-injection": "off", // Many errors, not big benefits
 		"@EslintSecurity/detect-non-literal-regexp": "error",
 		"@EslintSecurity/detect-bidi-characters": "error",
 
@@ -419,7 +423,7 @@ export default defineFlatConfig({
 		"@EslintSonar/no-all-duplicated-branches": "error",
 		"@EslintSonar/no-collapsible-if": "error",
 		"@EslintSonar/no-collection-size-mischeck": "error",
-		"@EslintSonar/no-duplicate-string": "error",
+		"@EslintSonar/no-duplicate-string": "off", // Bruh
 		"@EslintSonar/no-duplicated-branches": "error",
 		"@EslintSonar/no-element-overwrite": "error",
 		"@EslintSonar/no-empty-collection": "error",
@@ -456,12 +460,12 @@ export default defineFlatConfig({
 		// Imports
 		"@EslintImports/no-unresolved": "off", // [2, { caseSensitiveStrict: true }] Off cuz can have issues with bundlers
 		"@EslintImports/named": "off",
-		"@EslintImports/default": "warn",
+		"@EslintImports/default": "off", // ERRORS ["error"]
 		"@EslintImports/no-namespace": "off", // needed in some libs
 		"@EslintImports/export": "error",
 		"@EslintImports/exports-lats": "off",
 		// "@EslintImports/no-mutable-exports"
-		"@EslintImports/namespace": "error",
+		"@EslintImports/namespace": "off", // ERRORS ["error"]
 		"@EslintImports/no-mutable-exports": "error",
 		"@EslintImports/extensions": "off", // Don't want to configure
 		"@EslintImports/no-restricted-paths": "off",
@@ -502,7 +506,7 @@ export default defineFlatConfig({
 		"@EslintImports/no-import-module-exports": "off",
 		"@EslintImports/no-empty-named-blocks": "error",
 		"@EslintImports/exports-last": "off",
-		"@EslintImports/no-deprecated": "error",
+		"@EslintImports/no-deprecated": "off", // ERRORS ["error"]
 		"@EslintImports/no-named-as-default-member": "off",
 		"@EslintImports/no-amd": "off",
 		"@EslintPrefArrayFunc/from-map": "error",